Add UI interaction to the backend
Options are displayed to the user if requested by the backend,
and a progressbar is displayed while the backup/restore is done.

The frontend and backend are separated to make easier create alternative
implementations.

A property is_cancellable is added to the SectionView,
because the restore can't be cancelled, and need modify the toolbar
buttons according to that.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>