Replace combobox with palettes on Language cpsection
The Language-selection combobox does not work properly with touch and
the long list of language and country-variants is almost impossible to
use. (This is a revision of PR #279 -- as per the recommendation of
Gonzalo, no ellipsis or max character width is used since we have
ample room.)

This patch does three things:
 (1) separates language selection from country selection, thus making
     the language list less unwieldy; and
 (2) replaces the combobox with a palette of menus.
 (3) cleans up some outstanding issues with model.py

Try2: Includes pyflakes corrections; touch detection on label;
improvements to locale parsing (w/Gonzalo Odiard); and removal of special
case for Haitian Kreyol and Pashto since they have landed upstream.